# packages/api-client/openapi.json
{
  "openapi": "3.0.3",
  "info": {
    "title": "Workhorse API",
    "version": "0.1.0",
    "description": "Local-first runtime API for the Workhorse multi-workspace kanban."
  },
  "servers": [
    {
      "url": "http://127.0.0.1:3999"
    }
  ],
...
  "paths": {
...
    "/api/settings": {
      "get": {
        "operationId": "getSettings",
        "summary": "Read global settings",
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "Global settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          }
        }
      },
      "patch": {
        "operationId": "updateSettings",
        "summary": "Update global settings",
        "tags": [
          "Settings"
        ],
        "responses": {
          "200": {
            "description": "Updated global settings",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/SettingsResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateSettingsBody"
              }
            }
          }
        }
      }
    },
    "/api/workspaces": {
      "get": {
        "operationId": "listWorkspaces",
        "summary": "List registered workspaces",
        "tags": [
          "Workspaces"
        ],
        "responses": {
          "200": {
            "description": "Workspace collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspacesResponse"
                }
              }
            }
          }
        }
      },
      "post": {
        "operationId": "createWorkspace",
        "summary": "Register a workspace",
        "tags": [
          "Workspaces"
        ],
        "responses": {
          "201": {
            "description": "Created workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateWorkspaceBody"
              }
            }
          }
        }
      }
    },
    "/api/workspaces/pick-root": {
      "post": {
        "operationId": "pickWorkspaceRoot",
        "summary": "Open a local folder picker for a workspace root",
        "tags": [
          "Workspaces"
        ],
        "responses": {
          "200": {
            "description": "Selected workspace root path, or null when canceled",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PickWorkspaceRootResponse"
                }
              }
            }
          },
          "501": {
            "description": "Folder picker is unavailable on this platform",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        }
      }
    },
    "/api/workspaces/{workspaceId}": {
      "patch": {
        "operationId": "updateWorkspace",
        "summary": "Rename a workspace",
        "tags": [
          "Workspaces"
        ],
        "responses": {
          "200": {
            "description": "Updated workspace",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateWorkspaceBody"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteWorkspace",
        "summary": "Delete a workspace",
        "tags": [
          "Workspaces"
        ],
        "responses": {
          "200": {
            "description": "Deleted workspace id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteWorkspaceResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/workspaces/{workspaceId}/git/refs": {
      "get": {
        "operationId": "listWorkspaceGitRefs",
        "summary": "List Git refs for a workspace",
        "tags": [
          "Workspaces"
        ],
        "responses": {
          "200": {
            "description": "Workspace Git refs",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceGitRefsResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/workspaces/{workspaceId}/git/status": {
      "get": {
        "operationId": "getWorkspaceGitStatus",
        "summary": "Get Git sync status for a workspace",
        "tags": [
          "Workspaces"
        ],
        "responses": {
          "200": {
            "description": "Workspace Git sync status",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceGitStatusResponse"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/workspaces/{workspaceId}/git/pull": {
      "post": {
        "operationId": "pullWorkspace",
        "summary": "Pull latest changes from origin for a workspace",
        "tags": [
          "Workspaces"
        ],
        "responses": {
          "200": {
            "description": "Pull result",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/WorkspaceGitPullResponse"
                }
              }
            }
          },
          "400": {
            "description": "Pull failed",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Workspace not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "workspaceId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
    "/api/tasks": {
      "get": {
        "operationId": "listTasks",
        "summary": "List tasks across workspaces",
        "tags": [
          "Tasks"
        ],
        "responses": {
          "200": {
            "description": "Task collection",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TasksResponse"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "workspaceId",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string"
            }
          }
        ]
      },
      "post": {
        "operationId": "createTask",
        "summary": "Create a task card",
        "tags": [
          "Tasks"
        ],
        "responses": {
          "201": {
            "description": "Created task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateTaskBody"
              }
            }
          }
        }
      }
    },
    "/api/tasks/{taskId}": {
      "patch": {
        "operationId": "updateTask",
        "summary": "Update a task card",
        "tags": [
          "Tasks"
        ],
        "responses": {
          "200": {
            "description": "Updated task",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TaskResponse"
                }
              }
            }
          },
          "400": {
            "description": "Validation error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateTaskBody"
              }
            }
          }
        }
      },
      "delete": {
        "operationId": "deleteTask",
        "summary": "Delete a task card",
        "tags": [
          "Tasks"
        ],
        "responses": {
          "200": {
            "description": "Deleted task id",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/DeleteTaskResponse"
                }
              }
            }
          },
          "404": {
            "description": "Task not found",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ApiError"
                }
              }
            }
          }
        },
        "parameters": [
          {
            "name": "taskId",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string"
            }
          }
        ]
      }
    },
...
  },
  "components": {
    "schemas": {
      "ApiError": {
        "type": "object",
        "properties": {
          "ok": {
            "enum": [
              false
            ]
          },
          "error": {
            "type": "object",
            "properties": {
              "code": {
                "type": "string"
              },
              "message": {
                "type": "string"
              },
              "details": {
                "oneOf": [
                  {
                    "type": "string"
                  },
                  {
                    "type": "array",
                    "items": {
                      "$ref": "#/components/schemas/ApiErrorIssue"
                    }
                  }
                ]
              }
            },
            "required": [
              "code",
              "message"
            ]
          }
        },
        "required": [
          "ok",
          "error"
        ]
      },
...
      "SettingsResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/SettingsData"
          }
        },
        "required": [
          "ok",
          "data"
        ]
      },
...
      "WorkspacesResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/ListWorkspacesData"
          }
        },
        "required": [
          "ok",
          "data"
        ]
      },
...
      "WorkspaceResponse": {
        "type": "object",
        "properties": {
          "ok": {
            "enum": [
              true
            ]
          },
          "data": {
            "$ref": "#/components/schemas/WorkspaceData"
          }
        },
        "required": [
          "ok",
          "data"
        ]
      },
...
    }
  }
}